Cast to u32

+++ openssl/src/x509/mod.rs
@@ -399,8 +399,8 @@ impl X509Ref {
     }
 
     /// Returns the hash of the certificates subject
-    pub fn subject_name_hash(&self) -> u64 {
-        unsafe { ffi::X509_subject_name_hash(self.as_ptr()) }
+    pub fn subject_name_hash(&self) -> u32 {
+        unsafe { ffi::X509_subject_name_hash(self.as_ptr()) as u32 }
     }
 
     /// Returns this certificate's issuer name.
